finish LogIn SignIn menu
