+++ .idea/misc.xml
@@ -3,6 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="app/src/main/res/layout/activity_log_in_sign_in_menu.xml" value="0.25022665457842247" />
         <entry key="app/src/main/res/layout/activity_login_in_sign_in_menu.xml" value="0.25" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.33" />
       </map>
