--- app/src/main/java/com/bouboulecorp/thurii/LogInSignInMenu.kt
@@ -1,18 +1,0 @@
-package com.bouboulecorp.thurii
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.View
-
-class LogInSignInMenu : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_log_in_sign_in_menu)
-    }
-
-    fun goToLoginPage(view: View) {
-        val intent = Intent(this, MainActivity::class.java)
-        startActivity(intent)
-    }
-}
