+++ app/src/main/java/com/bouboulecorp/thurii/activities/LogIn/LoginActivity.kt
@@ -1,0 +1,12 @@
+package com.bouboulecorp.thurii.activities.LogIn
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.bouboulecorp.thurii.R
+
+class LoginActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login)
+    }
+}
