+++ app/src/main/res/layout/activity_log_in_sign_in_menu.xml
@@ -4,17 +4,49 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LogInSignInMenu">
+    android:background="@color/custom_color_background"
+    tools:context=".activities.LogInSignInMenu.LogInSignInMenuActivity">
+
+    <ImageView
+        android:id="@+id/imageView2"
+        android:layout_width="171dp"
+        android:layout_height="72dp"
+        android:layout_marginBottom="28dp"
+        app:layout_constraintBottom_toTopOf="@+id/log_in_menu_button"
+        app:layout_constraintStart_toStartOf="@+id/log_in_menu_button"
+        app:srcCompat="@drawable/thurii_with_slogan_flou" />
 
     <Button
-        android:id="@+id/login_menu_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
-        android:layout_marginTop="492dp"
+        android:id="@+id/log_in_menu_button"
+        style="@style/ButtonGreenStyle"
         android:onClick="goToLoginPage"
-        android:text="Button"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:text="@string/log_in_button"
+        android:layout_marginBottom="28dp"
+        app:layout_constraintBottom_toTopOf="@+id/sign_in_menu_button"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <Button
+        android:id="@+id/sign_in_menu_button"
+        style="@style/ButtonStroke"
+        android:onClick="goToLoginPage"
+        android:text="@string/sign_in_button"
+        android:layout_marginBottom="107dp"
+        app:layout_constraintBottom_toTopOf="@+id/textView"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"/>
+
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="323dp"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="25dp"
+        android:text="@string/test_link"
+        android:textAlignment="center"
+        android:textColor="@color/white"
+        android:textColorLink="@color/custom_color_primary_"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
