+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,17 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+
+    <!--Custom Colors-->
+
+    <color name="custom_color_primary_">#41BB99</color>
+    <color name="custom_color_secondary_btn">#A616191E</color>
+    <color name="custom_color_text_primary">#EBF5F9</color>
+    <color name="custom_color_text_secondary">#B3EBF5F9</color>
+    <color name="custom_color_background">#0B0F0F</color>
+    <color name="custom_color_slider_widget_selected">#99EBF5F9</color>
+    <color name="custom_color_slider_widget_unselected">#33EBF5F9</color>
+    <color name="custom_color_border_btn">#AEEBF5F9</color>
+    <color name="custom_color_separate_line">#181A1B</color>
+    <color name="gradient_black_filter_button">#16191E</color>
 </resources>
