+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,9 @@
 <resources>
     <string name="app_name">Thurii</string>
+
+
+    <!-- Login SignIn Menu -->
+    <string name="log_in_button">S\'inscrire</string>
+    <string name="sign_in_button">J\'ai déjà un compte</string>
+    <string name="test_link">En continuant, vous acceptez nos <a href="https://youtu.be/dQw4w9WgXcQ">Conditions d\'utilisation </a>et <a href="https://youtu.be/dQw4w9WgXcQ">Privacy Policy</a></string>
 </resources>
