+++ app/src/main/res/values/themes.xml
@@ -10,7 +10,73 @@
         <item name="colorSecondaryVariant">@color/teal_700</item>
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
+        <item name="android:statusBarColor">?attr/colorOnPrimary</item>
+        <item name="android:windowLightStatusBar">true</item>
         <!-- Customize your theme here. -->
+    </style>
+
+    <!-- Button Style. -->
+
+    <!-- Button Style with stroke. -->
+    <style name="ButtonStroke">
+        <item name="android:layout_width">334dp</item>
+        <item name="android:layout_height">55dp</item>
+        <item name="android:paddingStart">40dp</item>
+        <item name="android:paddingEnd">40dp</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textSize">15sp</item>
+        <item name="android:textStyle">bold</item>
+        <item name="android:textColor">@color/custom_color_text_primary</item>
+        <item name="backgroundTint">@color/custom_color_secondary_btn</item>
+        <item name="cornerRadius">8dp</item>
+        <item name="layout_goneMarginBottom">104dp</item>
+        <item name="strokeColor">@color/custom_color_border_btn</item>
+        <item name="strokeWidth">2dp</item>
+    </style>
+
+    <!-- Button Style without stroke and dark bg. -->
+
+    <style name="ButtonWithoutStrokeDarkBg">
+        <item name="android:layout_width">334dp</item>
+        <item name="android:layout_height">55dp</item>
+        <item name="android:paddingStart">40dp</item>
+        <item name="android:paddingEnd">40dp</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textSize">15sp</item>
+        <item name="android:textStyle">bold</item>
+        <item name="android:textColor">@color/custom_color_text_primary</item>
+        <item name="backgroundTint">@color/custom_color_secondary_btn</item>
+        <item name="cornerRadius">8dp</item>
+        <item name="layout_goneMarginBottom">104dp</item>
+    </style>
+
+    <!-- Button Green Style. -->
+    <style name="ButtonGreenStyle">
+        <item name="android:layout_width">334dp</item>
+        <item name="android:layout_height">55dp</item>
+        <item name="android:paddingStart">40dp</item>
+        <item name="android:paddingEnd">40dp</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textSize">15sp</item>
+        <item name="android:textStyle">bold</item>
+        <item name="android:textColor">@color/custom_color_text_primary</item>
+        <item name="backgroundTint">@color/custom_color_primary_</item>
+        <item name="cornerRadius">8dp</item>
+        <item name="layout_goneMarginBottom">104dp</item>
+    </style>
+
+    <!-- Button no clickable Style. -->
+    <style name="ButtonGreyStyle">
+        <item name="android:layout_width">334dp</item>
+        <item name="android:layout_height">55dp</item>
+        <item name="android:paddingStart">40dp</item>
+        <item name="android:paddingEnd">40dp</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textSize">15sp</item>
+        <item name="android:textStyle">bold</item>
+        <item name="android:textColor">@color/custom_color_text_secondary</item>
+        <item name="backgroundTint">@color/custom_color_primary_</item>
+        <item name="cornerRadius">8dp</item>
+        <item name="layout_goneMarginBottom">104dp</item>
     </style>
 </resources>
